requirements.txt
fastapi
uvicorn
pydantic
supabase
google-generativeai
python-dotenv
httpx
websockets
